Renamed test helpers as a spec file so it is not included in the prod build
